+++ blueprints/flexberry-core/files/__root__/models/custom-inflector-rules.js
@@ -1,0 +1,5 @@
+import Inflector from 'ember-inflector';
+
+const inflector = Inflector.inflector;
+
+<%= inflectorIrregular %>